+++ src/TFM.h
@@ -4,7 +4,7 @@
 ** This file is part of dvisvgm -- the DVI to SVG converter             **
 ** Copyright (C) 2005-2010 Martin Gieseking <martin.gieseking@uos.de>   **
 **                                                                      **
-** This program is free software; you can redistribute it and/or        ** 
+** This program is free software; you can redistribute it and/or        **
 ** modify it under the terms of the GNU General Public License as       **
 ** published by the Free Software Foundation; either version 3 of       **
 ** the License, or (at your option) any later version.                  **
@@ -15,7 +15,7 @@
 ** GNU General Public License for more details.                         **
 **                                                                      **
 ** You should have received a copy of the GNU General Public License    **
-** along with this program; if not, see <http://www.gnu.org/licenses/>. ** 
+** along with this program; if not, see <http://www.gnu.org/licenses/>. **
 *************************************************************************/
 
 #ifndef TFM_H
@@ -38,7 +38,7 @@ class TFM
 		double getCharHeight (int c) const;
 		double getCharDepth (int c) const;
 		double getItalicCorr (int c) const;
-		UInt16 getChecksum () const {return _checksum;}
+		UInt32 getChecksum () const {return _checksum;}
 		UInt16 firstChar () const   {return _firstChar;}
 		UInt16 lastChar () const    {return _lastChar;}
 
@@ -46,7 +46,7 @@ class TFM
 		bool readFromStream (std::istream &is);
 
    private:
-		UInt16 _checksum;
+		UInt32 _checksum;
 		UInt16 _firstChar, _lastChar;
 		FixWord _designSize;  ///< design size of the font in TeX points (7227 pt = 254 cm)
 		std::vector<UInt32>  _charInfoTable;
